Cast delays to double because this is what _delay_us wants

+++ code/main.c
@@ -212,8 +212,8 @@ void beep(float freq, float duration, int on_position) {
   }
 
   float wavelength = (1 / freq) * 1000;
-  uint32_t cycles = duration / wavelength;
-  uint16_t half_period_us = (uint16_t)1000 * wavelength / 2;
+  uint32_t cycles = (uint32_t)duration / wavelength;
+  double half_period_us = (double)1000 * wavelength / 2;
 
   for (uint32_t i = 0; i < cycles; i++) {
     _delay_us(half_period_us);
